server/requirements.txt
alabaster==0.7.8
anyio==3.3.0
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
Babel==2.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
bleach==3.3.1
blinker==1.4
blis==0.7.4
breezy==3.0.2
Brlapi==0.7.0
catalogue==2.0.4
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cymem==2.0.5
dbus-python==1.2.16
debugpy==1.4.1
decorator==5.0.9
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.7
distlib==0.3.2
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2.12
django-cors-headers==3.10.0
djangorestframework==3.12.4
docopt==0.6.2
docutils==0.16
dulwich==0.19.15
duplicity==0.8.12.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.1.0/en_core_web_lg-3.1.0-py3-none-any.whl
entrypoints==0.3
fasteners==0.14.1
fastimport==0.9.8
feedparser==5.2.1
filelock==3.0.12
future==0.18.2
gitsome==0.8.0
gpg===1.13.1-unknown
httplib2==0.14.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.10.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.10.1
jupyterlab==3.1.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib-inline==0.1.2
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
murmurhash==1.0.5
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
nltk==3.6.2
notebook==6.4.0
numpy==1.21.1
numpydoc==0.7.0
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pandas==1.3.1
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.2
pathy==0.6.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==2.0.2
ply==3.11
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==2.0.10
protobuf==3.6.1
ptyprocess==0.7.0
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
pydantic==1.8.2
pygccxml==2.2.1
PyGithub==1.43.7
Pygments==2.9.0
PyGObject==3.36.0
pygraphviz==1.5
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.18.0
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-gitlab==2.0.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.1.0
regex==2021.7.6
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
scikit-learn==0.24.2
scipy==1.7.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.7.1
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
smart-open==5.1.0
sniffio==1.2.0
spacy==3.1.1
spacy-legacy==3.0.8
Sphinx==1.8.5
sqlparse==0.2.4
srsly==2.4.1
systemd-python==234
terminado==0.10.1
testpath==0.5.0
thinc==8.0.8
threadpoolctl==2.2.0
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.10.0.0
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.6.0
wadllib==1.3.3
wasabi==0.8.2
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==1.1.0
wrapt==1.11.2
xkit==0.0.0
xonsh==0.9.13
zipp==1.0.0
